Move range learn logic to RangeTeacher. Create higher order componet RangeTrainer to manage range state. Add RangeAdder scaffold
