--- RangeTrainer/ActionButtonGroup.css
@@ -1,4 +1,0 @@
-.actionButtonGroup {
-  display: flex;
-  justify-content: space-between;
-}
