+++ RangeTrainer/RangeAdder/RangeAdder.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import { View } from "react-native";
+
+class RangeAdder extends Component {
+  render() {
+    return <View />;
+  }
+}
+
+export default RangeAdder;
