+++ RangeTrainer/RangeTeacher/POSITIONS.js
@@ -1,0 +1,3 @@
+const POSITIONS = ["EP", "MP", "HJ", "CO", "BTN"];
+
+export default POSITIONS;
